requirements.txt
asttokens==3.0.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
dash==2.18.2
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.13
decorator==5.2.1
executing==2.2.0
Flask==3.0.3
geopandas==1.0.1
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.5
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.6.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.6.0
narwhals==1.26.0
nest-asyncio==1.6.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
platformdirs==4.3.7
plotly==6.0.0
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.1
pyogrio==0.10.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==310
pywin32-ctypes==0.2.3
pyzmq==26.3.0
requests==2.32.3
retrying==1.3.4
setuptools==75.8.0
shapely==2.0.7
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.0.6
zipp==3.21.0
